Refactor: 불필요한 콘솔 제거, feedback 컴포넌트 분리
